WIP: Profile Historical accuracy charts
Add a section below the "view history" button for "My Accuracy".

For each of the following charts it should only pull data for the current "challenge" (iteration)

The section should start with a radial area graph for the User's overall average accuracy percentage over for each type of shot - use the same colors as in the shooting session charts.

Below the area graph should be a section with the user's TargetAccuracyVisualizers for each type of shot - (it's okay to just use like 40 shot "hit/dots" to represent their rough accuracy for these ones) the percentage should be accurate for all of their accuracy data recorded in the current challenge/iteration.

Those TargetAccuracyVisualizers should be tappable to swap between the graphs of each shot type below...

Below that will be a line graph that visualizes the user's accuracy over time for the current challenge, it should show an additional grey "area line" to display what direction the player/user's accuracy is trending in (up or down and with how much variance).

All of the charts should use the same color scheme as elsewhere to represent each shot type, and make sure things are properly labelled.